Remove TODO to reverse the output of PKCS12_parse.
OpenSSL classified their behavior as a bug and are fixing it for the
next release. In principle it'd be more compatible to emulate OpenSSL's
bug and undo it when we update OPENSSL_VERSION_NUMBER, but use of
PKCS12_parse is rare and this behavior is confusing, so let's leave it
as-is.

Bug: 250
Change-Id: I5f9825490a8afde67272dfaf476b35dbde94b59c