Generalize make_errors.go to allow EVP covering multiple directories.
In doing so, this switches make_errors.go to take library names as
parameters rather than detecting it from the CWD. (I considered
detecting it, but then we'd need to map evp -> crypto/whatever and
crypto/whatever -> evp in both directions.)

Since crypto/hpke currently sits in the EVP namespace, I've gone ahead
and added that, so it should be easier to define new errors in
crypto/hpke. I've not added crypto/cipher, etc., yet moving those will
be a breaking change.

Bug: 398
Change-Id: Ibae2afd46e076891fa517c377b540b2e492516f0