runner: Fix writeClientHash and writeRecord ordering.
Per the comment in writeClientHash, we should writeClientHash before
writeRecord to get the sequence numbers right. Some of the client HRR
bits are still wrong, but I'll fix those as part of tidying up the HRR
path in a later commit.

(This doesn't actually matter because only DTLS uses sequence numbers,
and we don't support DTLS 1.3.)

Change-Id: I4cbc671f524d56c7f970b5ec0bceeb2641625d15