runner: Don't use the buffer in TLS 1.3.
All the comments say the buffer is only needed in TLS 1.2, but this
doesn't match the code. The code uses the buffer in one place, for ECH,
to avoid copying a hash.Hash. Go does support this, albeit in a *very*
roundabout way.

This is ugly but means we can now properly drop the handshake buffer in
TLS 1.3.

Change-Id: I4a1559a64fcb98ccfbab54de99402fe6f62725a1