Don't use SHA256(ticket) as the signaling session ID for tickets.
We've inherited some behavior from OpenSSL where, in ticket-based client
sessions, we fill in a placeholder session ID of SHA256(ticket). This
was done to avoid confusing other code in OpenSSL (and possibly
callers?) that assumed session_id_length != 0 determined validity.

Separately, TLS 1.2 session tickets are syntactically weird. The client
generates a fake signaling session ID, which the server echoes on
resumption. We end up sending the placeholder SHA256 value as this
signaling one. Since we already have code to generate random session IDs
anyway, use that instead to minimize unnecessary implementation quirks
visible on the wire.

This removes one of the places we still rely on this machinery.

Change-Id: I0de2781da72e2bbc030505611589c853f105ce9d