DTLS-SRTP is only defined for DTLS.
This avoids needing to worry about the interaction with renegotiation
which, in turn, means we can drop the init callback. (If we did support
DTLS renegotiation, we'd probably want to forbid the parameter from
changing anyway. Changing your SRTP parameters partway through will
likely confuse the RTP half of the application anyway.)

Change-Id: Ifef1e9479d9df296b69b0d296f6bef57b13da68e