runner: Implement ECH server for testing.
This implements draft-ietf-tls-esni-10.

This will be used to test the client implementation. While I'm here,
I've switched the setup logic in the server tests to use the new
ServerECHConfig type. I'll probably need to patch in various features
later for testing, but this should be a usable starting point.

Based on an initial implementation by Dan McArdle in
https://boringssl-review.googlesource.com/c/boringssl/+/46786

Bug: 275
Change-Id: I69523cda70c3da2ae505bcab837fd358195fb9e9