Add util/fetch_ech_config_list.go
I wrote this tool to make it easier to test the ECH client against
real-world servers with the bssl client tool. I found that manually
extracting an ECHConfigList from a raw HTTPS record is unnecessarily
painful.

The tool queries DNS over UDP for HTTPS records. If it finds an HTTPS
record in the response, it attempts to extract an ECHConfigList from the
"ech" SvcParam. It can save the extracted ECHConfigList to a file. Once
the ECH client implementation lands, the bssl client tool should have a
new flag that that takes the path to an ECHConfigList file.

I am using golang.org/x/net/dns/dnsmessage to parse the DNS response. I
recently added the |UnknownResource| type to this library to enable
callers (like us) to extract the bytes of otherwise-unsupported records
(like HTTPS). I updated the dependency with `go get -u golang.org/x/net`.

Although the bssl client tool knows how to resolve the address of its
"-connect" parameter, it is difficult to query HTTPS records in a
platform-agnostic way. If we decide the bssl client should directly
query HTTPS rather than leaning on fetch_ech_config_list.go, we should
look into libresolv. Specifically, the |res_query| function enables the
caller to query arbitrary record types. This may open its own can of
cross-platform worms; macOS and Linux typically ship with different
implementations and it is not available on Windows. For more info, see
`man 3 resolver`.

Bug: 275
Change-Id: I705591658921f60a958164a18b68ffb697c2ea4b