Add util/fetch_ech_config_list.go
I wrote this tool to make it easier to test ECH against real-world
servers with the bssl client tool. I found manually parsing the
ECHConfigList out of a serialized HTTPS record is unnecessarily painful.

The tool queries DNS over UDP for HTTPS records. If any HTTPS records
are present in the response, it attempts to extract the "ech" SvcParam,
which we expect to contain a serialized ECHConfigList.

It uses golang.org/x/net/dns/dnsmessage to parse the DNS response. I
recently added the |UnknownResource| type to this library, which now
enables us to extract the bytes of otherwise-unsupported records like
HTTPS. I updated the dependency with `go get -u golang.org/x/net`.

Although the bssl client tool knows how to resolve the address of its
"-connect" parameter, it is difficult to query newer record types like
HTTPS in a platform-agnostic way. If we decide the bssl client should
make its own DNS queries for HTTPS records, we should look into the
libresolv library's |res_query| function, which enables the caller to
specify the desired type. This may open its own can of cross-platform
worms; at a minimum, macOS and Linux typically ship with different
implementations and it is not available on Windows. For more info, see
`man 3 resolver`.

Bug: 275
Change-Id: I705591658921f60a958164a18b68ffb697c2ea4b