Add an option to permute ClientHello extension order.
Although not permitted by the TLS specification, systems sometimes
make assumptions about TLS extension order, or byte offsets of various
fields. This is a problem for TLS extensibility, so, to keep the
ecosystem healthy, add an API to reorder ClientHello extensions.

Since ECH, HelloRetryRequest, and HelloVerifyRequest are sensitive to
extension order, I've implemented this by per-connection permutation of
the indices in the kExtensions structure. This ensures that all
ClientHellos within a connection are consistently ordered. As follow-up
work, permuting the other messages would also be nice, though any server
messages would need to be incorporated in handshake hints.

Change-Id: I18ce39b4df5ee376c654943f07ec26a50e0923a9