Only clear not_resumable after the handshake.
In renegotiation handshakes and, later, ECH ClientHelloOuter handshakes,
we don't want to add sessions to the session cache. We also don't want
to release a session as resumable until the handshake completes.

But we're stuck with APIs like SSL_get_session working in the handshake,
so SSL_SESSION also needs to work as an in-progress bundle of connection
properties. We manage this with a not_resumable flag that's only cleared
after the handshake is done and the SSL_SESSION finalized.

Except TLS 1.2 ticket renewal currently clears it early. This won't
actually affect renegotiation or ClientHelloOuter because those will
never resume. Nonetheless, maintain the invariant by instead setting it
on hs->new_session. Note this does sacrifice a different invariant, that
ssl->session and hs->new_session are never set at the same time, but it
seems to be fine. This also means ssl_update_cache does not need to
special-case ticket renewal.

Change-Id: I03230cd9c63e5bee6bd60cd05c0439e16533c6d4